re #100 Added a list of assignments to the street view Time:4h0m
re #100 Added a popup window when you click an assignment, this window is animated. Time:4h0m
re #100 Added code behind the yes and no buttons on the assignment window, yes shows the guid of the room to be opened, no animates the closing of the window. Time: 4h30m